Cambios pins

+++ src/pins/pins.repository.ts
@@ -2,16 +2,20 @@ import { Pin } from "./entities/pins.entity";
 import {  In, Repository } from "typeorm";
 import { InjectRepository } from "@nestjs/typeorm";
 import { pinsDto, updateDto } from "./pinsDtos/pins.dto";
-import { BadRequestException, NotFoundException } from "@nestjs/common";
-import { CreateLikeDto } from "./pinsDtos/like.dto";
+import { BadRequestException, ForbiddenException, NotFoundException } from "@nestjs/common";
 import { Like } from "./entities/likes.entity";
 import { Comment } from "./entities/comments.entity";
 import { CommentDto } from "./pinsDtos/comments.dto";
 import { User } from "src/users/entities/user.entity";
 import { Hashtag } from "./entities/hashtag.entity";
 import { Category } from "../categories/category.entity";
+import { View } from "./entities/view.entity";
+import { Save } from "./entities/save.entity";
 
 export class PinsRepository {
+    
+    
+    
     constructor(
         @InjectRepository(Category)
         private readonly categoryRepo: Repository<Category>,
@@ -30,21 +34,43 @@ export class PinsRepository {
 
         @InjectRepository(Hashtag)
         private readonly hashtagRepo: Repository<Hashtag>,
+
+        @InjectRepository(View)
+        private readonly viewRepo: Repository<View>,
+
+        @InjectRepository(Save)
+        private readonly saveRepo: Repository<Save>
     ){}
 
-    async getPins(): Promise<Pin[]> {
-        return await this.pinsRepo.find()
+    // Create Query PINS Repository
+    
+    async createSearch(query: string) {
+        
+        return this.pinsRepo
+        .createQueryBuilder("p")
+        .leftJoinAndSelect("p.hashtags", "h")
+        .where("p.description ILIKE :q", {q: `%${query}%`})
+        .orWhere("h.tag ILIKE :q", { q: `%${query}%` })
+        .getMany()
+    }
+
+    // Create PINS Repository
+
+    async getPins(page: number, limit: number): Promise<Pin[]> {
+        return await this.pinsRepo.find({
+            skip: (page - 1) * 10,
+            take: limit,
+            order: {createdAt: "DESC"}
+        })
     }
 
     async pinsId(id: string): Promise<Pin | null>{
         return await this.pinsRepo.findOne({where: {id: id}})
     }
 
+    async createPins(dtoPin: pinsDto, idUser:string) {
 
-
-    async createPins(dtoPin: pinsDto, idCategory:string, idUser:string) {
-
-        const initialización = await this.categoryRepo.findOne({where: {id: idCategory}})
+        const initialización = await this.categoryRepo.findOne({where: {id: dtoPin.categoryId}})
         if(!initialización)throw new NotFoundException("Error initializing category.")
 
         const users = await this.userRepo.findOne({where: {id: idUser}})     
@@ -67,34 +93,49 @@ export class PinsRepository {
             like: create.likesCount,
             comment: create.commentsCount,
             view: create.views,
-            hashtag: create.hashtags
+            hashtag: create.hashtags,
+            date: create.createdAt
             }
     }
 
-    async modifiPins(dtoPin: updateDto, userId: string, hashtagId:string[]): Promise<Pin> {
-        const pin = await this.pinsRepo.findOne({where: {id: userId}})
+    async modifiPins(userId: string, dtoPin: updateDto, pinsId: string, hashtags: { id: string; tag: string }[]) {
+        
+        const user = await this.userRepo.findOne({where: {id: userId}})
+        if(!user) throw new NotFoundException("Error to modify the post.")
+
+        const pin = await this.pinsRepo.findOne({where: {id: pinsId}, relations: ["hashtags"]})
         if(!pin) throw new NotFoundException("Error to modify the post.")
 
-        const hashtag = await this.hashtagRepo.findOne({where: {id: In(hashtagId)}})
-        if(!hashtag) throw new NotFoundException("Hashtag not found.")
+         const updatedHashtags: Hashtag[] = []
+            for (const h of hashtags) {
+            const hashtag = await this.hashtagRepo.findOne({ where: { id: h.id } })
+            if (!hashtag) throw new NotFoundException(`Hashtag with id not found.`)
+
+                hashtag.tag = h.tag
+            updatedHashtags.push(await this.hashtagRepo.save(hashtag))
+            }
 
         const modifi =  this.pinsRepo.merge(
             pin, {    
                 ...dtoPin,
-                hashtags: [hashtag]
+                hashtags: updatedHashtags
             })
 
         return await this.pinsRepo.save(modifi)
     }
 
-
-    async deletePins(id: string): Promise<Pin> {
-        const pin = await this.pinsRepo.findOneBy({id: id})
-
+    async deletePins(id: string, userId: string): Promise<Pin> {
+            
+        const pin = await this.pinsRepo.findOne({where: {id: id}, relations: ["user"]})
         if(!pin) throw new NotFoundException("Error to delete the post.")
+
+        
+        if(pin.user.id !== userId) throw new ForbiddenException("You are not allowed to delete this post.")
 
         return await this.pinsRepo.remove(pin)   
     }
+
+    // Create Like PINS Repository
 
     async createLike(idPin:string, idUser: string): Promise<Like | { message: string;}> {
         const pin = await this.pinsRepo.findOne({where: { id: idPin}})
@@ -114,17 +155,24 @@ export class PinsRepository {
 
     }
 
-    async deleteLike(id: string): Promise<Like> {
+    async deleteLike(id: string, userId:string): Promise<Like> {
+        
+        
         const remove = await this.likeRepo.findOne({where: {
             id: id,},
-        relations:["pin"]
+        relations:["pin", "user"]
         })
         
         if(!remove) throw new NotFoundException("Post not found.")
+
+        if(remove.user.id !== userId) throw new ForbiddenException("You are not allowed to delete this like.")
+
         
         await this.pinsRepo.decrement({id: remove.pin.id}, "likesCount", 1)
         return await this.likeRepo.remove(remove)
     }
+
+    // Create Comment PINS Repository
 
     async createComment(userId: string, pinId:string , comment: CommentDto) {
         const pin = await this.pinsRepo.findOne({where: {id: pinId}})
@@ -150,28 +198,96 @@ export class PinsRepository {
         }
     }
 
-    async modifieComment(id:string, comment: CommentDto): Promise<Comment> {
-        const commentId = await this.commentRepo.findOne({where: {id: id}})
+    async modifieComment(id: string, comment: CommentDto, userId: string): Promise<Comment> {
+        const commentId = await this.commentRepo.findOne({where: {id: id}, relations: ["user"]})
         if(!commentId) throw new NotFoundException("Comment not found.")
+        if(commentId.user.id !== userId) throw new ForbiddenException("You are not allowed to modifie this comment.")  
 
         const modifiComment = this.commentRepo.merge(commentId, comment)
         return await this.commentRepo.save(modifiComment)
     }
 
-    async deleteComment(id: string): Promise<Comment> {
-        const commentId = await this.commentRepo.findOne({where: {id: id}}) 
+    async deleteComment(id: string, userId: string): Promise<Comment> {
+        const commentId = await this.commentRepo.findOne({where: {id: id}, relations:["user"]}) 
         if(!commentId) throw new NotFoundException("Comment not found.")
+        if(commentId.user.id !== userId) throw new ForbiddenException("You are not allowed to delete this comment.")
 
         return await this.commentRepo.remove(commentId)
     }
 
-    async createSearch(query: string) {
+    // Create View PINS Repository
+
+    async createView(idUser: string, idPins: string) {
+        const pin = await this.pinsRepo.findOne({where: {id: idPins}})
+        if(!pin) throw new NotFoundException("Post not found.")
         
-        return this.pinsRepo
-        .createQueryBuilder("p")
-        .leftJoinAndSelect("p.hashtags", "h")
-        .where("p.description ILIKE :q", {q: `%${query}%`})
-        .orWhere("h.tag ILIKE :q", { q: `%${query}%` })
-        .getMany()
+        const user = await this.userRepo.findOne({where: {id: idUser}})
+        if(!user) throw new NotFoundException("User not found.")
+
+        const viewCreate = this.viewRepo.create({
+            user: {id: user.id},
+            pin: {id: pin.id}
+        })
+
+        await this.pinsRepo.increment({id: pin.id}, "viewsCount", 1)
+        await this.viewRepo.save(viewCreate)
+
+        return viewCreate;
     }
+
+    // Create Save PINS Repository
+
+    async createGetSave(idUser:string) {
+        const user = await this.userRepo.findOne({ where: { id: idUser } });
+        if (!user) throw new NotFoundException("User not found.");
+
+        
+
+        const save = await this.saveRepo.find({
+            where: {
+                user: {id: user.id}},
+                relations: ["pin"]
+        })
+
+       const pins = await save.map(e=> e.pin)
+
+       return pins
+
+    }
+
+    async createSave(idPin: string, idUser: string ) {
+        
+        const pin = await this.pinsRepo.findOne({ where: { id: idPin } });
+        if (!pin) throw new NotFoundException("Post not found.");
+
+        const user = await this.userRepo.findOne({ where: { id: idUser } });
+        if (!user) throw new NotFoundException("User not found.");
+
+
+        const existing = await this.saveRepo.findOne({
+        where: { user: { id: user.id }, pin: { id: pin.id } },
+        });
+        if (existing) throw new BadRequestException("This post is already saved.");
+
+  
+        const save = this.saveRepo.create({
+            user: {id: user.id},
+            pin,
+        });
+
+        return await this.saveRepo.save(save);
+    }
+
+
+    async createDeleteSave(id: string, idUser: string) {
+        const deleteSave = await this.saveRepo.findOne({where: {id: id}, relations: ["user"]})
+        if(!deleteSave) throw new NotFoundException("Item not found.")
+
+        if(deleteSave.user.id !== idUser)throw new ForbiddenException("You are not allowed to delete this comment.")
+
+        await this.saveRepo.remove(deleteSave)    
+    }
+
+
+
 }
